fix: move HOT observer to useWalletAgnosticSignMessage
